Suggested fixes part 1
