--- src/components/Header.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react'
-import Navbar from './Navbar'
-
-const Header: React.FC = () => <Navbar />
-
-export default Header
