+++ src/components/navbar/Navbar.tsx
@@ -1,0 +1,68 @@
+import { CloseIcon, HamburgerIcon, MoonIcon, SunIcon } from '@chakra-ui/icons'
+import { Box, Collapse, Flex, IconButton, Link, useColorMode, useColorModeValue, useDisclosure } from '@chakra-ui/react'
+import * as React from 'react'
+import Logo from '../../assets/images/logo-with-name.svg'
+import DesktopNav from './DesktopNav'
+import MobileNav from './MobileNav'
+
+const WithSubnavigation: React.FC = () => {
+  const { isOpen, onToggle } = useDisclosure()
+  const { colorMode, toggleColorMode } = useColorMode()
+
+  return (
+    <Box align="center">
+      <Flex
+        bg={useColorModeValue('white', 'gray.800')}
+        color={useColorModeValue('gray.800', 'white')}
+        minH={{ base: '3rem', md: '4.5rem' }}
+        maxW={['100%', '100%', '720px', '1280px']}
+        py={{ base: 2 }}
+        px={{ base: 4 }}
+        align="center"
+      >
+        <Flex flex={{ base: 1, md: '1' }} ml={{ base: -2, md: 0 }} display={{ base: 'flex', md: 'none' }}>
+          <IconButton
+            onClick={onToggle}
+            icon={isOpen ? <CloseIcon w={3} h={3} /> : <HamburgerIcon w={5} h={5} />}
+            variant="ghost"
+            aria-label="Navigáció megnyitása"
+          />
+        </Flex>
+        <Flex flex={{ base: 1 }} justify={{ base: 'center', md: 'start' }}>
+          <Link
+            href="/"
+            display="block"
+            rounded="md"
+            _hover={{
+              textDecoration: 'none',
+              color: 'orange.500'
+            }}
+          >
+            <Box textAlign={{ base: 'center', md: 'left' }} mx="2">
+              <Logo style={{ fill: useColorModeValue('#000', '#fff') }} />
+            </Box>
+          </Link>
+        </Flex>
+        <Flex display={{ base: 'none', md: 'flex' }} flex={{ base: 1 }} justify={{ base: 'center', md: 'flex-end' }}>
+          <Flex display={{ base: 'none', md: 'flex' }} mx={4}>
+            <DesktopNav />
+          </Flex>
+        </Flex>
+        <Flex flex={{ base: 1, md: 0 }} mr={{ base: -2, md: 0 }} justify="flex-end">
+          <IconButton
+            aria-label="Sötét-világos mód váltás"
+            icon={colorMode === 'dark' ? <SunIcon w={5} h={5} /> : <MoonIcon w={5} h={5} />}
+            onClick={toggleColorMode}
+            variant="ghost"
+          />
+        </Flex>
+      </Flex>
+
+      <Collapse in={isOpen} animateOpacity>
+        <MobileNav />
+      </Collapse>
+    </Box>
+  )
+}
+
+export default WithSubnavigation
