+++ src/typings.d.ts
@@ -13,3 +13,8 @@ declare module '*.module.css' {
   const cssModule: CSSModule
   export = cssModule
 }
+
+declare module '*.svg' {
+  const content: any
+  export default content
+}
